+++ deepctr_torch/models/afm.py
@@ -7,19 +7,16 @@ Reference:
     (https://arxiv.org/abs/1708.04617)
 """
 import torch
-import torch.nn.functional as F
 
 from .basemodel import BaseModel
 from ..layers import FM, AFMLayer
 
 
 class AFM(BaseModel):
-
     """Instantiates the Attentional Factorization Machine architecture.
 
     :param linear_feature_columns: An iterable containing all the features used by linear part of the model.
     :param dnn_feature_columns: An iterable containing all the features used by deep part of the model.
-    :param embedding_size: positive integer,sparse feature embedding_size
     :param use_attention: bool,whether use attention or not,if set to ``False``.it is the same as **standard Factorization Machine**
     :param attention_factor: positive integer,units in attention net
     :param l2_reg_linear: float. L2 regularizer strength applied to linear part
@@ -34,21 +31,21 @@ class AFM(BaseModel):
 
     """
 
-    def __init__(self,linear_feature_columns, dnn_feature_columns, embedding_size=8, use_attention=True, attention_factor=8,
+    def __init__(self, linear_feature_columns, dnn_feature_columns, use_attention=True, attention_factor=8,
                  l2_reg_linear=1e-5, l2_reg_embedding=1e-5, l2_reg_att=1e-5, afm_dropout=0, init_std=0.0001, seed=1024,
                  task='binary', device='cpu'):
-        super(AFM, self).__init__(linear_feature_columns, dnn_feature_columns, embedding_size=embedding_size,
+        super(AFM, self).__init__(linear_feature_columns, dnn_feature_columns,
                                   dnn_hidden_units=[],
                                   l2_reg_linear=l2_reg_linear,
                                   l2_reg_embedding=l2_reg_embedding, l2_reg_dnn=0, init_std=init_std,
                                   seed=seed,
-                                  dnn_dropout=0, dnn_activation=F.relu,
+                                  dnn_dropout=0, dnn_activation='relu',
                                   task=task, device=device)
 
         self.use_attention = use_attention
 
         if use_attention:
-            self.fm = AFMLayer(embedding_size, attention_factor, l2_reg_att, afm_dropout,
+            self.fm = AFMLayer(self.embedding_size, attention_factor, l2_reg_att, afm_dropout,
                                seed, device)
             self.add_regularization_loss(self.fm.attention_W, l2_reg_att)
         else:
@@ -59,7 +56,7 @@ class AFM(BaseModel):
     def forward(self, X):
 
         sparse_embedding_list, _ = self.input_from_feature_columns(X, self.dnn_feature_columns,
-                                                                                  self.embedding_dict,support_dense=False)
+                                                                   self.embedding_dict, support_dense=False)
         logit = self.linear_model(X)
         if len(sparse_embedding_list) > 0:
             if self.use_attention: