+++ deepctr_torch/models/dcn.py
@@ -7,18 +7,17 @@ Reference:
 """
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
 
 from .basemodel import BaseModel
 from ..inputs import combined_dnn_input
 from ..layers import CrossNet, DNN
 
-class DCN(BaseModel):
 
+class DCN(BaseModel):
     """Instantiates the Deep&Cross Network architecture.
 
+    :param linear_feature_columns: An iterable containing all the features used by linear part of the model.
     :param dnn_feature_columns: An iterable containing all the features used by deep part of the model.
-    :param embedding_size: positive int or str,sparse feature embedding_size.If set to "auto",it will be 6*pow(cardinality,025)
     :param cross_num: positive integet,cross layer number
     :param dnn_hidden_units: list,list of positive integer or empty list, the layer number and units in each layer of DNN
     :param l2_reg_embedding: float. L2 regularizer strength applied to embedding vector
@@ -35,16 +34,15 @@ class DCN(BaseModel):
     
     """
 
-    def __init__(self,
-                 dnn_feature_columns, embedding_size=8, cross_num=2,
+    def __init__(self,linear_feature_columns,
+                 dnn_feature_columns, cross_num=2,
                  dnn_hidden_units=(128, 128), l2_reg_linear=0.00001,
                  l2_reg_embedding=0.00001, l2_reg_cross=0.00001, l2_reg_dnn=0, init_std=0.0001, seed=1024,
                  dnn_dropout=0,
-                 dnn_activation=F.relu, dnn_use_bn=False, task='binary', device='cpu'):
+                 dnn_activation='relu', dnn_use_bn=False, task='binary', device='cpu'):
 
-        super(DCN, self).__init__(linear_feature_columns=[],
+        super(DCN, self).__init__(linear_feature_columns=linear_feature_columns,
                                   dnn_feature_columns=dnn_feature_columns,
-                                  embedding_size=embedding_size,
                                   dnn_hidden_units=dnn_hidden_units,
                                   l2_reg_embedding=l2_reg_embedding, l2_reg_dnn=l2_reg_dnn, init_std=init_std,
                                   seed=seed,
@@ -52,19 +50,19 @@ class DCN(BaseModel):
                                   task=task, device=device)
         self.dnn_hidden_units = dnn_hidden_units
         self.cross_num = cross_num
-        self.dnn = DNN(self.compute_input_dim(dnn_feature_columns, embedding_size), dnn_hidden_units,
+        self.dnn = DNN(self.compute_input_dim(dnn_feature_columns), dnn_hidden_units,
                        activation=dnn_activation, use_bn=dnn_use_bn, l2_reg=l2_reg_dnn, dropout_rate=dnn_dropout,
                        init_std=init_std, device=device)
         if len(self.dnn_hidden_units) > 0 and self.cross_num > 0:
-            dnn_linear_in_feature = self.compute_input_dim(dnn_feature_columns, embedding_size) + dnn_hidden_units[-1]
+            dnn_linear_in_feature = self.compute_input_dim(dnn_feature_columns) + dnn_hidden_units[-1]
         elif len(self.dnn_hidden_units) > 0:
             dnn_linear_in_feature = dnn_hidden_units[-1]
         elif self.cross_num > 0:
-            dnn_linear_in_feature = self.compute_input_dim(dnn_feature_columns, embedding_size)
+            dnn_linear_in_feature = self.compute_input_dim(dnn_feature_columns)
 
         self.dnn_linear = nn.Linear(dnn_linear_in_feature, 1, bias=False).to(
             device)
-        self.crossnet = CrossNet(in_features=self.compute_input_dim(dnn_feature_columns, embedding_size),
+        self.crossnet = CrossNet(in_features=self.compute_input_dim(dnn_feature_columns),
                                  layer_num=cross_num, seed=1024, device=device)
         self.add_regularization_loss(
             filter(lambda x: 'weight' in x[0] and 'bn' not in x[0], self.dnn.named_parameters()), l2_reg_dnn)
@@ -74,6 +72,7 @@ class DCN(BaseModel):
 
     def forward(self, X):
 
+        logit = self.linear_model(X)
         sparse_embedding_list, dense_value_list = self.input_from_feature_columns(X, self.dnn_feature_columns,
                                                                                   self.embedding_dict)
 
@@ -83,15 +82,14 @@ class DCN(BaseModel):
             deep_out = self.dnn(dnn_input)
             cross_out = self.crossnet(dnn_input)
             stack_out = torch.cat((cross_out, deep_out), dim=-1)
-            final_logit = self.dnn_linear(stack_out)
+            logit += self.dnn_linear(stack_out)
         elif len(self.dnn_hidden_units) > 0:  # Only Deep
             deep_out = self.dnn(dnn_input)
-            final_logit = self.dnn_linear(deep_out)
+            logit += self.dnn_linear(deep_out)
         elif self.cross_num > 0:  # Only Cross
             cross_out = self.crossnet(dnn_input)
-            final_logit = self.dnn_linear(cross_out)
+            logit += self.dnn_linear(cross_out)
         else:  # Error
-            raise NotImplementedError
-
-        y_pred = self.out(final_logit)
+            pass
+        y_pred = self.out(logit)
         return y_pred