+++ tests/models/WDL_test.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 import pytest
+
 from deepctr_torch.models import WDL
-from ..utils import get_test_data, SAMPLE_SIZE, check_model
+from ..utils import get_test_data, SAMPLE_SIZE, check_model, get_device
 
 
 @pytest.mark.parametrize(
@@ -13,10 +14,10 @@ def test_WDL(sparse_feature_num, dense_feature_num):
     model_name = "WDL"
     sample_size = SAMPLE_SIZE
     x, y, feature_columns = get_test_data(
-        sample_size, sparse_feature_num, dense_feature_num)
+        sample_size, sparse_feature_num=sparse_feature_num, dense_feature_num=dense_feature_num)
 
     model = WDL(feature_columns, feature_columns,
-                dnn_hidden_units=[32, 32], dnn_dropout=0.5)
+                dnn_hidden_units=[32, 32], dnn_dropout=0.5, device=get_device())
     check_model(model, model_name, x, y)